requirements.txt
streamlit==0.82.0
pytesseract==0.3.8
Pillow==8.0.1
gTTS==2.2.2

